refactor(runtimeprovider): simplify graph and resolver boundaries
